requirements.txt
gunicorn
streamlit
agno
python-dotenv
firecrawl
pydantic
fastapi
uvicorn
google-genai
openai